Albergue_mascotas/requirements.txt
altgraph==0.17.2
apsw==3.28.0.post1
apturl==0.5.2
asgiref==3.5.0
asttokens==2.0.5
attrs==19.3.0
autopep8==1.6.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-admin-interface==0.19.0
django-colorfield==0.6.3
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
django-jet==1.0.8
django-multiselectfield==0.1.12
django-phonenumber-field==6.1.0
django-suit==0.2.28
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
fasteners==0.14.1
feedparser==5.2.1
future==0.18.2
gunicorn==20.1.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
importlib-metadata==1.5.0
imutils==0.5.4
ipykernel==6.9.1
ipython==8.1.1
ipython_genutils==0.2.0
jedi==0.18.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
mahotas==1.4.12
Mako==1.1.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.0
matplotlib==3.1.2
matplotlib-inline==0.1.3
mechanize==0.4.5
monotonic==1.5
more-itertools==4.2.0
mpmath==1.1.0
msgpack==0.6.2
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.4
nose==1.3.7
numpy==1.17.4
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==20.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
psycogreen==1.0.1
psycopg2==2.8.4
pure-eval==0.2.2
PyAudio==0.2.11
pycairo==1.16.2
pychm==0.8.6
pycodestyle==2.8.0
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.36.0
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==2.4.6
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
regex==2019.8.19
reportlab==3.5.34
repoze.lru==0.7
requests==2.22.0
requests-unixsocket==0.2.0
Routes==2.4.1
scipy==1.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
sqlparse==0.4.2
ssh-import-id==5.10
stack-data==0.2.0
sympy==1.5.1
system-service==0.3
systemd-python==234
tablib==3.2.1
thrift==0.13.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
urwid==2.0.1
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.5
X Tile==3.3
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zeroconf==0.24.4
zipp==1.0.0
